Fast-math flag on fabs is unsupported yet
