+++ llvm_util/known_fns.cpp
@@ -120,7 +120,8 @@ known_call(llvm::CallInst &i, const llvm::TargetLibraryInfo &TLI,
   case llvm::LibFunc_fabs:
   case llvm::LibFunc_fabsf: {
     RETURN_KNOWN(
-      make_unique<UnaryOp>(*ty, value_name(i), *args[0], UnaryOp::FAbs));
+      make_unique<UnaryOp>(*ty, value_name(i), *args[0], UnaryOp::FAbs,
+                           parse_fmath(i)));
   }
   default:
     RETURN_FAIL_KNOWN();
